+++ Sources/GrAIdient/Layer2D/Transform/Rotate2D.swift
@@ -470,7 +470,7 @@ public class Rotate2D: Layer2D
             let pNbChannels: [UInt32] = [UInt32(nbChannels)]
             let pNbBatch: [UInt32] = [UInt32(batchSize)]
             let pDimensions: [UInt32] = [UInt32(width), UInt32(height)]
-            let pAngle: [Float] = [Float16(_angle)]
+            let pAngle: [Float] = [Float(_angle)]
             
             command = MetalKernel.get.createCommand(
                 "rotate2DBackward", deviceID: deviceID
